server/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.1.1
requests==2.32.5
